Minor cleanup, post-refactoring of cmd
