+++ cmd/nerdctl/main.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"os"
 	"runtime"
-	"strconv"
 	"strings"
 
 	"github.com/fatih/color"
@@ -166,15 +165,15 @@ func initRootCmdFlags(rootCmd *cobra.Command, tomlPath string) (*pflag.FlagSet, 
 	rootCmd.PersistentFlags().Bool("debug", cfg.Debug, "debug mode")
 	rootCmd.PersistentFlags().Bool("debug-full", cfg.DebugFull, "debug mode (with full output)")
 	// -a is aliases (conflicts with nerdctl images -a)
-	AddPersistentStringFlag(rootCmd, "address", []string{"a", "H"}, nil, []string{"host"}, aliasToBeInherited, cfg.Address, "CONTAINERD_ADDRESS", `containerd address, optionally with "unix://" prefix`)
+	helpers.AddPersistentStringFlag(rootCmd, "address", []string{"a", "H"}, nil, []string{"host"}, aliasToBeInherited, cfg.Address, "CONTAINERD_ADDRESS", `containerd address, optionally with "unix://" prefix`)
 	// -n is aliases (conflicts with nerdctl logs -n)
-	AddPersistentStringFlag(rootCmd, "namespace", []string{"n"}, nil, nil, aliasToBeInherited, cfg.Namespace, "CONTAINERD_NAMESPACE", `containerd namespace, such as "moby" for Docker, "k8s.io" for Kubernetes`)
+	helpers.AddPersistentStringFlag(rootCmd, "namespace", []string{"n"}, nil, nil, aliasToBeInherited, cfg.Namespace, "CONTAINERD_NAMESPACE", `containerd namespace, such as "moby" for Docker, "k8s.io" for Kubernetes`)
 	rootCmd.RegisterFlagCompletionFunc("namespace", completion.NamespaceNames)
-	AddPersistentStringFlag(rootCmd, "snapshotter", nil, nil, []string{"storage-driver"}, aliasToBeInherited, cfg.Snapshotter, "CONTAINERD_SNAPSHOTTER", "containerd snapshotter")
+	helpers.AddPersistentStringFlag(rootCmd, "snapshotter", nil, nil, []string{"storage-driver"}, aliasToBeInherited, cfg.Snapshotter, "CONTAINERD_SNAPSHOTTER", "containerd snapshotter")
 	rootCmd.RegisterFlagCompletionFunc("snapshotter", completion.SnapshotterNames)
 	rootCmd.RegisterFlagCompletionFunc("storage-driver", completion.SnapshotterNames)
-	AddPersistentStringFlag(rootCmd, "cni-path", nil, nil, nil, aliasToBeInherited, cfg.CNIPath, "CNI_PATH", "cni plugins binary directory")
-	AddPersistentStringFlag(rootCmd, "cni-netconfpath", nil, nil, nil, aliasToBeInherited, cfg.CNINetConfPath, "NETCONFPATH", "cni config directory")
+	helpers.AddPersistentStringFlag(rootCmd, "cni-path", nil, nil, nil, aliasToBeInherited, cfg.CNIPath, "CNI_PATH", "cni plugins binary directory")
+	helpers.AddPersistentStringFlag(rootCmd, "cni-netconfpath", nil, nil, nil, aliasToBeInherited, cfg.CNINetConfPath, "NETCONFPATH", "cni config directory")
 	rootCmd.PersistentFlags().String("data-root", cfg.DataRoot, "Root directory of persistent nerdctl state (managed by nerdctl, not by containerd)")
 	rootCmd.PersistentFlags().String("cgroup-manager", cfg.CgroupManager, `Cgroup manager to use ("cgroupfs"|"systemd")`)
 	rootCmd.RegisterFlagCompletionFunc("cgroup-manager", completion.CgroupManagerNames)
@@ -182,8 +181,8 @@ func initRootCmdFlags(rootCmd *cobra.Command, tomlPath string) (*pflag.FlagSet, 
 	// hosts-dir is defined as StringSlice, not StringArray, to allow specifying "--hosts-dir=/etc/containerd/certs.d,/etc/docker/certs.d"
 	rootCmd.PersistentFlags().StringSlice("hosts-dir", cfg.HostsDir, "A directory that contains <HOST:PORT>/hosts.toml (containerd style) or <HOST:PORT>/{ca.cert, cert.pem, key.pem} (docker style)")
 	// Experimental enable experimental feature, see in https://github.com/containerd/nerdctl/blob/main/docs/experimental.md
-	AddPersistentBoolFlag(rootCmd, "experimental", nil, nil, cfg.Experimental, "NERDCTL_EXPERIMENTAL", "Control experimental: https://github.com/containerd/nerdctl/blob/main/docs/experimental.md")
-	AddPersistentStringFlag(rootCmd, "host-gateway-ip", nil, nil, nil, aliasToBeInherited, cfg.HostGatewayIP, "NERDCTL_HOST_GATEWAY_IP", "IP address that the special 'host-gateway' string in --add-host resolves to. Defaults to the IP address of the host. It has no effect without setting --add-host")
+	helpers.AddPersistentBoolFlag(rootCmd, "experimental", nil, nil, cfg.Experimental, "NERDCTL_EXPERIMENTAL", "Control experimental: https://github.com/containerd/nerdctl/blob/main/docs/experimental.md")
+	helpers.AddPersistentStringFlag(rootCmd, "host-gateway-ip", nil, nil, nil, aliasToBeInherited, cfg.HostGatewayIP, "NERDCTL_HOST_GATEWAY_IP", "IP address that the special 'host-gateway' string in --add-host resolves to. Defaults to the IP address of the host. It has no effect without setting --add-host")
 	return aliasToBeInherited, nil
 }
 
@@ -332,92 +331,4 @@ Config file ($NERDCTL_TOML): %s
 		subCmd.InheritedFlags().AddFlagSet(aliasToBeInherited)
 	}
 	return rootCmd, nil
-}
-
-// AddPersistentStringFlag is similar to AddStringFlag but persistent.
-// See https://github.com/spf13/cobra/blob/main/user_guide.md#persistent-flags to learn what is "persistent".
-func AddPersistentStringFlag(cmd *cobra.Command, name string, aliases, localAliases, persistentAliases []string, aliasToBeInherited *pflag.FlagSet, value string, env, usage string) {
-	if env != "" {
-		usage = fmt.Sprintf("%s [$%s]", usage, env)
-	}
-	if envV, ok := os.LookupEnv(env); ok {
-		value = envV
-	}
-	aliasesUsage := fmt.Sprintf("Alias of --%s", name)
-	p := new(string)
-
-	// flags is full set of flag(s)
-	// flags can redefine alias already used in subcommands
-	flags := cmd.Flags()
-	for _, a := range aliases {
-		if len(a) == 1 {
-			// pflag doesn't support short-only flags, so we have to register long one as well here
-			flags.StringVarP(p, a, a, value, aliasesUsage)
-		} else {
-			flags.StringVar(p, a, value, aliasesUsage)
-		}
-		// non-persistent flags are not added to the InheritedFlags, so we should add them manually
-		f := flags.Lookup(a)
-		aliasToBeInherited.AddFlag(f)
-	}
-
-	// localFlags are local to the rootCmd
-	localFlags := cmd.LocalFlags()
-	for _, a := range localAliases {
-		if len(a) == 1 {
-			// pflag doesn't support short-only flags, so we have to register long one as well here
-			localFlags.StringVarP(p, a, a, value, aliasesUsage)
-		} else {
-			localFlags.StringVar(p, a, value, aliasesUsage)
-		}
-	}
-
-	// persistentFlags cannot redefine alias already used in subcommands
-	persistentFlags := cmd.PersistentFlags()
-	persistentFlags.StringVar(p, name, value, usage)
-	for _, a := range persistentAliases {
-		if len(a) == 1 {
-			// pflag doesn't support short-only flags, so we have to register long one as well here
-			persistentFlags.StringVarP(p, a, a, value, aliasesUsage)
-		} else {
-			persistentFlags.StringVar(p, a, value, aliasesUsage)
-		}
-	}
-}
-
-// AddPersistentBoolFlag is similar to AddBoolFlag but persistent.
-// See https://github.com/spf13/cobra/blob/main/user_guide.md#persistent-flags to learn what is "persistent".
-func AddPersistentBoolFlag(cmd *cobra.Command, name string, aliases, nonPersistentAliases []string, value bool, env, usage string) {
-	if env != "" {
-		usage = fmt.Sprintf("%s [$%s]", usage, env)
-	}
-	if envV, ok := os.LookupEnv(env); ok {
-		var err error
-		value, err = strconv.ParseBool(envV)
-		if err != nil {
-			log.L.WithError(err).Warnf("Invalid boolean value for `%s`", env)
-		}
-	}
-	aliasesUsage := fmt.Sprintf("Alias of --%s", name)
-	p := new(bool)
-	flags := cmd.Flags()
-	for _, a := range nonPersistentAliases {
-		if len(a) == 1 {
-			// pflag doesn't support short-only flags, so we have to register long one as well here
-			flags.BoolVarP(p, a, a, value, aliasesUsage)
-		} else {
-			flags.BoolVar(p, a, value, aliasesUsage)
-		}
-	}
-
-	persistentFlags := cmd.PersistentFlags()
-	persistentFlags.BoolVar(p, name, value, usage)
-	for _, a := range aliases {
-		if len(a) == 1 {
-			// pflag doesn't support short-only flags, so we have to register long one as well here
-			persistentFlags.BoolVarP(p, a, a, value, aliasesUsage)
-		} else {
-			persistentFlags.BoolVar(p, a, value, aliasesUsage)
-		}
-	}
 }
